Sprite ordering for overlap
Completely removed the dependency of the main sprite to be defined on
the top of the stack, and the tileset to be defined on the bottom. Every
frame, sort the sprite stack by y order, so that things that should be
in front of other things are rendered on top.