Add external scene loading support
Scenes are now defined by external json files loaded at runtime,
rather than by javascript constructs in the engine source file. This
helps to restrict the initial load time, and to make it easier to
export scenes from external editors.

Tile collision is really finnicky unless the area and carea are the
same size. This is now implemented, and a matching tileset has been
added.

A bug that causes save data to apply to the wrong sprites is now fixed.