Add screen.coffee to abstract DOM interaction
On the whole, the engine's graphical scope is limited to modifying its
own canvas. So, messing with CSS and other aspects of the DOM from
within that context feels a little weird. The screen class is meant to
contain all of the functions that work with the document. The sysinf
implementation is now located here, as well as CSS resizing,
fullscreen, and messaging.