Support range notation in external scene definitions
Range notation is useful for specifying many files with a single entry
in the resource list. Now, range notation is supported in the json
scene definition format, rather than just for internally defined scenes.

Also, a minor bug in the AI follow code, which causes the AI sprite to
not get out of the way properly in tight hallways, has been fixed.