Support for preloaded child nodes
Normally, an externally defined child node is only loaded when it's
needed, immediately before its own resources are loaded. This is
generally what you want, but for small nodes that don't require
resources of their own (such as, a short conversation with an NPC),
it's better to consider a node as nothing more than a resource of its
parent. To "preload" a node in this way, write a tilde (~) at the
beginning of its pathname in the child list.

It is now possible to write child lists in external node files, as a
list of strings under .k. Preloaded nodes here are also supported.