Move map collision data
Collision data in a tilemap (whether a tile is a wall or not) used to be
stored by changing the sign of a tile reference in the tilegraph to positive
or negative. This has been separated and moved. Rather than the collision
data being associated with the tile graph, it is now associated with the
tileset, which saves space and creates less room for bugs. Also, collision
with a tile is now represented by a number, rather than a sign, which makes
it possible to have more than two kinds of collision.