'bottom' variable removed from sprite
Sprites are rendered in order of increasing y-coordinate, so that sprites
below other sprites are rendered on top. This has caused problems with the
tile system in the past, because tiles used to change their y-coordinates
arbitrarily so collision would work. Because of this, a 'bottom' variable
was added so that tiles were always rendered first. Now that the collision
system has been replaced, and is no longer dependent on this arbitrary
change of position, the y-coordinate stays at 0, and so the 'bottom'
variable is no longer needed. This variable has now been removed.