Audio synced with input
The audio server has been on the top of the render stack, to keep it from
being blocked. This is problematic, though, because it means that all audio
control from each frame doesn't execute until the beginning of the following
frame, which causes a noticable sync issue. Now, the audio server runs after
the render stack. This way, it never gets blocked, but it also executes in
the proper frame.